strategies/requirements.txt
click
cryptography
matplotlib
empyrical
pandas
pyfolio
systematic-trading